09-API-working-example/API/requirements.txt
flask
psycopg2-binary
requests
ipykernel